Quick writeup of the planner.py wrapper python file
